Switched to JSON and included testapp
The plugin uses JSON now instead of XML.
A lot of people had problems with the XML data.
I also fixed the documentation and included the necessary
code for CSRF protection in the example.
Furthermore, a testapp was created to make testing
and developing easier. It's in a rough shape right now but better than nothing